+++ mycodo/mycodo_flask/routes_page.py
@@ -5,6 +5,7 @@ import datetime
 import glob
 import logging
 import resource
+import socket
 import subprocess
 import sys
 import time
@@ -18,6 +19,7 @@ from flask import flash
 from flask import redirect
 from flask import render_template
 from flask import request
+from flask import send_file
 from flask import url_for
 from flask.blueprints import Blueprint
 from flask_babel import gettext
@@ -36,6 +38,7 @@ from mycodo.config import INSTALL_DIRECTORY
 from mycodo.config import KEEPUP_LOG_FILE
 from mycodo.config import LOGIN_LOG_FILE
 from mycodo.config import MATH_INFO
+from mycodo.config import MYCODO_VERSION
 from mycodo.config import OUTPUTS
 from mycodo.config import OUTPUT_INFO
 from mycodo.config import RESTORE_LOG_FILE
@@ -258,6 +261,19 @@ def page_notes():
 
         if form_note_show.notes_show.data:
             notes = utils_notes.show_notes(form_note_show)
+        elif form_note_show.notes_export.data:
+            data = utils_notes.export_notes(form_note_show)
+            # Send zip file to user
+            return send_file(
+                data,
+                mimetype='application/zip',
+                as_attachment=True,
+                attachment_filename=
+                'Mycodo_Notes_{mv}_{host}_{dt}.zip'.format(
+                    mv=MYCODO_VERSION,
+                    host=socket.gethostname().replace(' ', ''),
+                    dt=datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S"))
+            )
         else:
             if form_tag_add.tag_add.data:
                 utils_notes.tag_add(form_tag_add)